Changes
So updated the features and now the attack pacman eat the pellets :). However they are greedy and keep them in their stomach instead of returning them to the base.
- Need to add a feature that tell its to return back to base after a certain amount of pellets are collected

Also I switched off Thompson Sampling as it did not seem to work well with approximate Q-learning and I am going for a greedy epsilon explore vs exploit function.